refactor seeding cli tool (extend functionality)
