+++ .PKGINFO
@@ -1,0 +1,11 @@
+# Generated by nfpm
+arch = x86_64
+builddate = 1774468219
+license = MIT
+packager = Unknown Packager
+pkgbase = supabase
+pkgdesc = Supabase CLI
+pkgname = supabase
+pkgver = 2.84.2-1
+size = 94924984
+url = https://supabase.com
