--- components/admin-views/account/Account.tsx
@@ -1,203 +1,0 @@
-'use client'
-
-import React, { useState, useEffect } from 'react'
-import { useRouter } from 'next/navigation'
-import { getAdminUser, updateAdminProfile, updateAdminPassword } from '@/actions/admin/account'
-
-interface AdminUser {
-  id: number
-  email: string
-  fullName?: string
-  role: 'admin' | 'user'
-}
-
-const AdminAccount: React.FC = () => {
-  const [user, setUser] = useState<AdminUser | null>(null)
-  const [loading, setLoading] = useState(true)
-  const [saving, setSaving] = useState(false)
-  const [error, setError] = useState('')
-  const [success, setSuccess] = useState(false)
-  const [fullName, setFullName] = useState('')
-  const [password, setPassword] = useState('')
-  const [confirmPassword, setConfirmPassword] = useState('')
-  const router = useRouter()
-
-  useEffect(() => {
-    const fetchUser = async () => {
-      const userData = await getAdminUser()
-      if (!userData) {
-        router.push('/admin/login')
-        return
-      }
-      setUser(userData)
-      setFullName(userData.fullName || '')
-      setLoading(false)
-    }
-
-    fetchUser()
-  }, [router])
-
-  const handleUpdateProfile = async (e: React.FormEvent) => {
-    e.preventDefault()
-    setError('')
-    setSuccess(false)
-    setSaving(true)
-
-    const result = await updateAdminProfile(fullName)
-
-    if (!result.ok) {
-      setError(result.error || 'حدث خطأ')
-    } else {
-      setSuccess(true)
-      setTimeout(() => setSuccess(false), 3000)
-    }
-    setSaving(false)
-  }
-
-  const handleChangePassword = async (e: React.FormEvent) => {
-    e.preventDefault()
-    setError('')
-    setSuccess(false)
-
-    if (password !== confirmPassword) {
-      setError('كلمات المرور غير متطابقة')
-      return
-    }
-
-    if (password.length < 8) {
-      setError('كلمة المرور يجب أن تكون 8 أحرف على الأقل')
-      return
-    }
-
-    setSaving(true)
-
-    const result = await updateAdminPassword(password)
-
-    if (!result.ok) {
-      setError(result.error || 'حدث خطأ')
-    } else {
-      setSuccess(true)
-      setPassword('')
-      setConfirmPassword('')
-      setTimeout(() => setSuccess(false), 3000)
-    }
-    setSaving(false)
-  }
-
-  if (loading) {
-    return (
-      <div className="min-h-screen flex items-center justify-center bg-gray-100">
-        <div className="text-center">جاري التحميل...</div>
-      </div>
-    )
-  }
-
-  return (
-    <div className="min-h-screen bg-gray-100 py-8">
-      <div className="max-w-2xl mx-auto">
-        <h1 className="text-2xl font-bold text-center mb-8">إعدادات الحساب</h1>
-
-        {error && (
-          <div className="bg-red-100 border border-red-400 text-red-700 px-4 py-3 rounded mb-4">
-            {error}
-          </div>
-        )}
-
-        {success && (
-          <div className="bg-green-100 border border-green-400 text-green-700 px-4 py-3 rounded mb-4">
-            تم التحديث بنجاح
-          </div>
-        )}
-
-        <div className="bg-white rounded-lg shadow-md p-6 mb-6">
-          <h2 className="text-xl font-bold mb-4">المعلومات الشخصية</h2>
-          <form onSubmit={handleUpdateProfile}>
-            <div className="mb-4">
-              <label className="block text-gray-700 text-sm font-bold mb-2">
-                البريد الإلكتروني
-              </label>
-              <input
-                type="email"
-                value={user?.email || ''}
-                disabled
-                className="w-full px-3 py-2 border border-gray-300 rounded-lg bg-gray-100"
-              />
-            </div>
-
-            <div className="mb-4">
-              <label className="block text-gray-700 text-sm font-bold mb-2">
-                الاسم الكامل
-              </label>
-              <input
-                type="text"
-                value={fullName}
-                onChange={(e) => setFullName(e.target.value)}
-                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-primary"
-              />
-            </div>
-
-            <div className="mb-4">
-              <label className="block text-gray-700 text-sm font-bold mb-2">
-                الدور
-              </label>
-              <input
-                type="text"
-                value={user?.role === 'admin' ? 'مسؤول' : 'مستخدم'}
-                disabled
-                className="w-full px-3 py-2 border border-gray-300 rounded-lg bg-gray-100"
-              />
-            </div>
-
-            <button
-              type="submit"
-              disabled={saving}
-              className="bg-primary text-white font-bold py-2 px-4 rounded-lg hover:bg-primary/90 disabled:opacity-50"
-            >
-              {saving ? 'جاري الحفظ...' : 'حفظ التغييرات'}
-            </button>
-          </form>
-        </div>
-
-        <div className="bg-white rounded-lg shadow-md p-6">
-          <h2 className="text-xl font-bold mb-4">تغيير كلمة المرور</h2>
-          <form onSubmit={handleChangePassword}>
-            <div className="mb-4">
-              <label className="block text-gray-700 text-sm font-bold mb-2">
-                كلمة المرور الجديدة
-              </label>
-              <input
-                type="password"
-                value={password}
-                onChange={(e) => setPassword(e.target.value)}
-                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-primary"
-                minLength={8}
-              />
-            </div>
-
-            <div className="mb-4">
-              <label className="block text-gray-700 text-sm font-bold mb-2">
-                تأكيد كلمة المرور الجديدة
-              </label>
-              <input
-                type="password"
-                value={confirmPassword}
-                onChange={(e) => setConfirmPassword(e.target.value)}
-                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-primary"
-              />
-            </div>
-
-            <button
-              type="submit"
-              disabled={saving || (!password && !confirmPassword)}
-              className="bg-primary text-white font-bold py-2 px-4 rounded-lg hover:bg-primary/90 disabled:opacity-50"
-            >
-              {saving ? 'جاري التحديث...' : 'تغيير كلمة المرور'}
-            </button>
-          </form>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-export default AdminAccount
