+++ payload.config.ts
@@ -32,13 +32,6 @@ export default buildConfig({
       createFirstUser: '/first-user',
       account: '/account',
     },
-    components: {
-      views: {
-        login: { Component: '@/components/admin-views/login/Login' },
-        firstUser: { Component: '@/components/admin-views/first-user/FirstUser' },
-        account: { Component: '@/components/admin-views/account/Account' },
-      },
-    },
   },
   collections: [
     User,
