+++ scripts/db/seed/collections/articles.ts
@@ -1,63 +1,6 @@
 import type { Payload } from 'payload'
-import { fakerAR as faker } from '@faker-js/faker'
-import { Article } from '@/payload-types'
-import { generateLexicalRichText } from './utils'
-import type { SeedOptions, SeedResult } from '../seed-config'
-import { withTransaction } from './with-transaction'
-
-const authors = [
-  'الشيخ محمد بن صالح العثيمين',
-  'الشيخ عبد العزيز بن عبد الله بن باز',
-  'الشيخ محمد ناصر الدين الألباني',
-  'الشيخ محمد بن إبراهيم آل الشيخ',
-  'الشيخ عبد الرحمن بن ناصر السعدي',
-  'الشيخ محمد بن صالح المنجد',
-  'الشيخ محمد بن صالح العثيمين',
-  'الشيخ عبد العزيز بن عبد الله بن باز',
-  'الشيخ محمد ناصر الدين الألباني',
-  'الشيخ محمد بن إبراهيم آل الشيخ',
-  'الشيخ عبد الرحمن بن ناصر السعدي',
-  'الشيخ محمد بن صالح المنجد',
-]
-
-const tags = [
-  'فقه',
-  'عقيدة',
-  'حديث',
-  'تفسير',
-  'سيرة',
-  'أخلاق',
-  'دعوة',
-  'فقه',
-  'عقيدة',
-  'حديث',
-  'تفسير',
-  'سيرة',
-  'أخلاق',
-  'دعوة',
-]
-
-const types: Article['type'][] = ['aqidah', 'fiqh', 'hadith', 'other']
-
-export const createArticle = async (payload: Payload, mediaIds: number[]) => {
-  const article = await payload.create({
-    collection: 'articles',
-    data: {
-      title: faker.lorem.words(3),
-      type: faker.helpers.arrayElement(types),
-      author: faker.helpers.arrayElement(authors),
-      description: faker.lorem.words(6),
-      publishDate: faker.date.past().toISOString(),
-      tags: faker.helpers
-        .arrayElements(tags, 3)
-        .map((tag) => ({ name: tag, id: faker.string.ulid() })),
-      image: faker.helpers.arrayElement(mediaIds),
-      content: generateLexicalRichText() as any,
-    },
-  })
-
-  return article
-}
+import { createArticle } from '../factories/article'
+import { withTransaction, type SeedOptions, type SeedResult } from '../shared/types'
 
 export const seedArticles = async (payload: Payload, options: SeedOptions = {}): Promise<SeedResult> => {
   const count = options.count ?? 50
