+++ scripts/db/seed/collections/index.ts
@@ -1,0 +1,8 @@
+export { seedUsers, createAdminUser } from './users'
+export { seedMedia } from './media'
+export { seedBooks } from './books'
+export { seedActivities } from './activities'
+export { seedArticles } from './articles'
+export { seedLoans } from './loans'
+export { seedBookFavorites } from './book-favorites'
+export { seedActivityRegistrations } from './activity-registrations'
