+++ scripts/db/seed/collections/loans.ts
@@ -1,31 +1,6 @@
 import type { Payload } from 'payload'
-import { fakerAR as faker } from '@faker-js/faker'
-import type { SeedOptions, SeedResult } from '../seed-config'
-import { withTransaction } from './with-transaction'
-
-export const createLoan = async (payload: Payload, bookIds: number[], userIds: number[]) => {
-  const statuses = ['pending', 'approved', 'returned', 'overdue'] as const
-  const status = faker.helpers.arrayElement(statuses)
-
-  const loanDate = faker.date.past({ years: 1 })
-  const dueDate = faker.date.future({ refDate: loanDate })
-  const returnDate =
-    status === 'returned' ? faker.date.between({ from: loanDate, to: dueDate }) : null
-
-  const loan = await payload.create({
-    collection: 'loans',
-    data: {
-      book: faker.helpers.arrayElement(bookIds),
-      user: faker.helpers.arrayElement(userIds),
-      status,
-      loanDate: new Date(loanDate).toISOString().split('T')[0],
-      dueDate: new Date(dueDate).toISOString().split('T')[0],
-      returnDate: returnDate ? new Date(returnDate).toISOString().split('T')[0] : null,
-    },
-  })
-
-  return loan
-}
+import { createLoan } from '../factories/loan'
+import { withTransaction, type SeedOptions, type SeedResult } from '../shared/types'
 
 export const seedLoans = async (payload: Payload, options: SeedOptions = {}): Promise<SeedResult> => {
   const count = options.count ?? 30
