+++ scripts/db/seed/factories/book/data.ts
@@ -1,0 +1,31 @@
+export const authors = [
+  'الشيخ محمد بن صالح العثيمين',
+  'الشيخ عبد العزيز بن عبد الله بن باز',
+  'الشيخ محمد ناصر الدين الألباني',
+  'الشيخ محمد بن إبراهيم آل الشيخ',
+  'الشيخ عبد الرحمن بن ناصر السعدي',
+  'الشيخ محمد بن صالح المنجد',
+  'الشيخ محمد بن صالح العثيمين',
+  'الشيخ عبد العزيز بن عبد الله بن باز',
+  'الشيخ محمد ناصر الدين الألباني',
+  'الشيخ محمد بن إبراهيم آل الشيخ',
+  'الشيخ عبد الرحمن بن ناصر السعدي',
+  'الشيخ محمد بن صالح المنجد',
+]
+
+export const tags = [
+  'فقه',
+  'عقيدة',
+  'حديث',
+  'تفسير',
+  'سيرة',
+  'أخلاق',
+  'دعوة',
+  'فقه',
+  'عقيدة',
+  'حديث',
+  'تفسير',
+  'سيرة',
+  'أخلاق',
+  'دعوة',
+]
