+++ scripts/db/seed/factories/index.ts
@@ -1,0 +1,8 @@
+export { createUser, createAdminUser } from './user'
+export { createMedia } from './media'
+export { createBook } from './book'
+export { createActivity } from './activity'
+export { createArticle } from './article'
+export { createLoan } from './loan'
+export { createBookFavorite } from './book-favorite'
+export { createActivityRegistration } from './activity-registration'
